Moved some stuff around, added higher level test and internal http client stuff
