+++ pipe-http-server-cloud/src/main/java/com/tesco/aqueduct/pipe/http/CloudPipeStateProvider.java
@@ -1,6 +1,7 @@
 package com.tesco.aqueduct.pipe.http;
 
 import com.tesco.aqueduct.pipe.api.MessageReader;
+import com.tesco.aqueduct.pipe.api.PipeStateResponse;
 
 import java.util.List;
